Keeping user login
Moved stack navigator which controls which screen to first show to a newly added entry point file
