+++ android/app/src/main/res/layout/launch_screen.xml
@@ -7,7 +7,8 @@
     android:gravity="center">
 
     <ImageView
-        android:layout_width="136dp"
-        android:layout_height="39dp"
+        android:layout_width="200dp"
+        android:layout_height="200dp"
+        android:layout_marginTop="24dp"
         android:src="@mipmap/icon" />
 </LinearLayout>
