+++ android/app/src/main/res/values/colors.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="blue">#4F6D7A</color>
+    <color name="primary_dark">#4F6D7A</color>
 </resources>
